+++ addons/es/verifactu/bill.go
@@ -54,6 +54,39 @@ func normalizeInvoice(inv *bill.Invoice) {
 			delete(inv.Tax.Ext, ExtKeyCorrectionType)
 		}
 	}
+
+	normalizeInvoicePartyIdentity(inv.Customer)
+}
+
+func normalizeInvoicePartyIdentity(cus *org.Party) {
+	if cus == nil {
+		return
+	}
+	if cus.TaxID != nil && cus.TaxID.Country == "ES" && cus.TaxID.Code != "" {
+		// Spanish NIFs are already handled
+		return
+	}
+	if len(cus.Identities) == 0 {
+		// nothing to do if no identities
+		return
+	}
+	id := cus.Identities[0]
+	var code cbc.Code
+	switch id.Key {
+	case org.IdentityKeyPassport:
+		code = ExtCodeIdentityTypePassport
+	case org.IdentityKeyForeign:
+		code = ExtCodeIdentityTypeForeign
+	case org.IdentityKeyResident:
+		code = ExtCodeIdentityTypeResident
+	case org.IdentityKeyOther:
+		code = ExtCodeIdentityTypeOther
+	}
+	if !code.IsEmpty() {
+		id.Ext = id.Ext.Merge(tax.Extensions{
+			ExtKeyIdentityType: code,
+		})
+	}
 }
 
 func validateInvoice(inv *bill.Invoice) error {
@@ -79,7 +112,10 @@ func validateInvoice(inv *bill.Invoice) error {
 			validation.Skip,
 		),
 		validation.Field(&inv.Notes,
-			org.ValidateNotesHasKey(org.NoteKeyGeneral),
+			validation.Each(
+				validation.By(validateNote),
+				validation.Skip,
+			),
 			validation.Skip,
 		),
 	)
@@ -173,4 +209,17 @@ func validateInvoicePreceding(inv *bill.Invoice) validation.RuleFunc {
 			),
 		)
 	}
+}
+
+func validateNote(val any) error {
+	note, ok := val.(*org.Note)
+	if !ok || note == nil || note.Key != org.NoteKeyGeneral {
+		return nil
+	}
+	return validation.ValidateStruct(note,
+		validation.Field(&note.Text,
+			validation.Length(0, 500),
+			validation.Skip,
+		),
+	)
 }